# faster circshift! for SparseMatrixCSC