--- a/stdlib/SparseArrays/test/sparse.jl
+++ b/stdlib/SparseArrays/test/sparse.jl
@@ -2408,6 +2408,17 @@ end
     @test oneunit(A) isa SparseMatrixCSC{Second}
     @test one(sprand(2, 2, 0.5)) isa SparseMatrixCSC{Float64}
     @test one(A) isa SparseMatrixCSC{Int}
+
+@testset "circshift" begin
+    A = sprand(40, 30, 0.3)
+    @test nnz(A) == nnz(circshift(A,(1,2)))
+    @test circshift(A, (3,4)) == circshift(Matrix(A), (3,4))
+    @test circshift(A, (7,-4)) == circshift(Matrix(A), (7,-4))
+    @test circshift(A, 14) == circshift(Matrix(A), 14)
+    O = similar(A)
+    circshift!(O, A, (11,13))
+    @test nnz(O) == nnz(A)
+    @test O == circshift(A, (11,13))
 end
 
 end # module