So this is another (much faster) version of the circshift! implementation for SparseMatrixCSC. Unfortunately, the code is much less legible than the two-liner of #30300, but

- avoids allocations ~~almost~~ completely
- has exactly two `mod` calls

In its current state, for SparseMatrixCSC matrices, it seems comparable to the dense version for small dense matrices, faster for larger ones and much much faster for both sparse and larger ones.

~~Question1: is it possible to replace the `O .= similar(X)` in the first line and just check/assert that allocated memory is compatible (i.e. colptr, rowval and nzval are of the same length and n,m are ==)?  I suppose no, because it would be a breaking change. This would avoid a double allocation in the call from circshift. Otherwise, I suppose I could move everything to a helper function and implement circshift instead of relying on the generic one. If this is the way to go, I would replace `O .= similar(X)` by just allocating the memory (no need to copy colptr and rowval)~~

EDIT: I realized that this question is kind of stupid... I just `resize!` colptr, rowval and nzval -- this does not involve allocation if they are already of the right size. It also makes things work when the output has different type than the input -- which didn't work before. I added some tests that would have catched it.

~~Question2: ~~even without `O .= similar(X)`,~~ there seems to be some very small allocation reported by `@benchmark`... I don't know where it comes from (but I may be missinterpreting).~~

I believe that this is a good improvement wrt to the current situation. The implementation is reasonably straightforward. Thoughts?

Updated benchmarks:

Summary of mean times, raw data [here](https://github.com/JuliaLang/julia/files/2660625/benchmarks.txt)



x                                    |  MASTER | BRANCH
--------------------------------|----------------|-----------------
sprand(10,10,1.0)          | 1.211 μs     | 1.218 μs
sprand(10,10,0.1)          | 789.932 ns | 765.545 ns
sprand(1000,1000,1.0)  | 71.192 ms  | 3.904 ms
sprand(1000,1000,0.1)  | 37.105 ms  | 196.435 μs
sprand(1000,1000,0.01)| 11.689 ms  | 26.761 μs


Updated benchmarks with some `@inbounds`, raw data [here](https://github.com/JuliaLang/julia/files/2663239/bench2.txt)


x                                    |  MASTER | BRANCH
--------------------------------|----------------|-----------------
sprand(10,10,1.0)          | 1.211 μs     | 1.172 μs
sprand(10,10,0.1)          | 789.932 ns | 725.866 ns
sprand(1000,1000,1.0)  | 71.192 ms  | 3.656 ms
sprand(1000,1000,0.1)  | 37.105 ms  | 167.782 μs
sprand(1000,1000,0.01)| 11.689 ms  | 20.948 μs


EDIT3: Found the solution to Question2 above: I think it the splat in the call to (dense) `circshift!` (in [multidimensional.jl):](https://github.com/JuliaLang/julia/blob/0a401f2b5dfa288e1812016b2b0311316de77697/base/multidimensional.jl#L935)

```
julia> x=rand(10); y=similar(x); @btime circshift!($y,$x,1);
  307.100 ns (3 allocations: 144 bytes)

julia> x=rand(10); y=similar(x); @btime circshift!($y,$x,(1,));
  35.272 ns (0 allocations: 0 bytes)

julia> @btime (1...,);
  263.556 ns (3 allocations: 144 bytes)
```
I thought that the temporary tuple would have been optimized out... can this be solved there? 

For the moment, I will just solve it localy. This gives a huge improvement for small matrices!

Updated benchmarks, raw data [here](https://github.com/JuliaLang/julia/files/2663513/bench3.txt)

x                                    |  MASTER | BRANCH
--------------------------------|----------------|-----------------
sprand(10,10,1.0)          | 1.211 μs     | 463.476 ns
sprand(10,10,0.1)          | 789.932 ns | 110.585 ns
sprand(1000,1000,1.0)  | 71.192 ms  | 3.669 ms
sprand(1000,1000,0.1)  | 37.105 ms  | 171.386 μs
sprand(1000,1000,0.01)| 11.689 ms  | 20.705 μs