@stevengj 2018-12-24
Are the 32-bit appveyor failures unrelated?