@abraunst 2018-12-24
> Are the 32-bit appveyor failures unrelated?

Seem unrelated to me.  OTOH, I have no idea of what they are :-)