@abraunst 2018-12-24
> Maybe rebase and force-push to see if the appveyor problem is something that was recently fixed.

Sure. Should I squash everything into a single commit?