@stevengj 2018-12-25
No need to squash — github allows us to squash when merging.  Thanks!